Author: Hynek Hanke <hanke@brailcom.org> --- log message follows this line --- Display of a custom role description in Login panel.
The role code from CMS is not passed to Wiking and the Wiking method User.role()
is canceled. Instead, there is a new method User.role_description() which returns
None by default but is overriden in CMS to return the role code description. The
option display_role_in_login_panel() is moved from CMS to Wiking.