Make sql_dir configuration option settable.
sql_dir is an application-specific (CMS) configuration option. Since
application-specific configuration is currently only possible through
explicitly enabling the option to be settable by Admin in WMI, and
sql_dir is not the kind of option Admin should deal with, but rather an
option the deployer should set it in config file, the option is moved
directly into Wiking's Configuration class, until application-specific
configuration through configuration files is implemented.