Avoid redirection to registration when the menu is empty.
Wiking performs redirection to the first menu item when the request goes to
server's root URI.  Since Wiking CMS adds some hidden menu items automatically,
they got used for this redirection when the application menu was actually
empty.  This is now avoided by omitting hidden menu items for this redirection.

Also, when the menu (non-hidden items) is empty and the user is not logged in,
Wiking will automatically ask for user's authentication in the presumption that
the menu will not be empty after authentication, which is the most common case
and which is typically more desired than returning Forbidden for the root page.