Add optional confirmation checkbox to the registration form.
Certain web sites may require some sort of confirmation of usage conditions or
a similar agreement.  The text of the agreement may be defined as a separate
page or a system text `cms.regintro' which appears above the registration form.
This commit makes it possible to add a reference to this text into the
registration form and a checkbox for the user to confirm.

If the new system text `cms.regconfirm' is defined (it is empty by default), it
is displayed within the registration form above the checkbox which allows the
user to confirm.  If the text is empty, the checkbox will not appear in the
form.  This commit doesn't yet allow the administrator to see whether the user
confirmed as it is not yet decided how it should be presented.  This should be
added soon.