Make it possible to set role containment also for automatic roles.
The "User Groups" module in WMI didn't contain roles, which are automatically
assigned to users, such as `ANYONE' or `AUTHENTICATED'.  But since it may be
desirable to manage the containment of automatic roles with other roles, the
module now contains all roles, including automatic roles.  As different kinds
of roles have different limitations, the user interface now describes the kind
of the role, related limitations and dynamically omits irrelevant controls.