Recognize role containment also for anonymous users.
When the role ANYONE contains other roles, authorization checking must consider
these roles when checking access of an anonymous user.  Anonymous user is a
member of the ANYONE role, thus he must be also member of its contained roles.

Role containment is resolved on construction of the `User' instance for logged
in users.  But there is no `User' instance for an anonymous user (`req.user()'
returns None).  Thus it is necessary to resolve the contained roles of the role
ANYONE during authorization checking.

This change required creation of a new metod `Application.contained_roles()'
because access checking is performed in `Request.check_roles()', which is part
of Wiking generic API and knows nothing about Wiking CMS `RoleSets' module.