Allow setting site_title from the configuration file.
Wiking CMS allows overriding configuraton options by values stored in a
database table.  But the site_title value was "not null" in the database so it
was not possible to omit this optin in the DB table and let it default to the
value from the config file.  As this may be sometimes useful, the not null
constraint is removed by this commit.