Confirmation of site conditions in Users layouts
Let users agree to the site conditions in the profile edit layout, if
they did not do so on registration. Display the status of confirmation
in the view and update layouts.

It would be desirable to show the confirmation status information
in a separate FieldSet in the view layout. However, in view layouts,
FieldSets containing only text content don't work.