Don't rely on `PytisModule._form()' returning a form instance
Some applications override _form() and wrap the form by additional content.
This is discouraged by the current docstring, but it is known (and also
documented at the end of the docstring), that some applications don't respect
it.  Thus calling form methods on the result of self._form() may lead to a
traceback when the result is not a form instance.  It is safer to query
`Form.heading_info()' directly within `PytisModule._form()'.

The docstring was improved to avoid confusion between the desired and the
current state.