Allow table data retrieval in JSON format
The new PytisModule action 'jsondata' allows retrieval of list form datain JSON
format.  The table is represented as a list of dictionaries (resp. Javascript
objects) where each dictionary corresponds to one table row and contains
key/value pairs for all columns visible in the form.  The values are not the
displayed form values, but rather the internal data values (such as key values
instead of "display" values for codebook columns).  The new action has not
assigned access rights by default, so they must be assigned by the application
if the action should be available for a particular form/module.