Get rid of `Request.uri_prefix()'
Multiple Wiking applications with differnt prefixes (the feature supposed to be
supported by this method) were actually never used in practice and this feature
would most likely not work as the prefixes are not respected at many places.
The method `Request.uri_prefix()' is kept for backwards compatibility, but
always returns an empty string now.  The removal of this feature is motivated
by the need for simplification of the Request API before merging the WSGI
implementation.