Factor out mod_python specific code to allow future mod_wsgi integration
This commit refactors the 'Request' and 'WikingRequest' class to split the web
server dependent methods into a separate class 'ServerInterface'.  The request
interface for wiking applications is built on top of the server interface.

The mod_python specific handler class and ServerInterface implementation is now
moved into a separate module 'mod_python_interface'.