Don't expect mod_python specific return values from request handling methods
The request handling methods were supposed to return req.done() when the
request was served (or 'Document' instance or a tuple (content_type, data)).
Since req.done() actually returned a mod_python specific value (zero as int),
we don't want this dependency.  This commit makes sure that int is never
expected by wiking and that the mod_python specific handler takes care of
returning the appropriate value to mod_python.