Page footers simplified
Various proclamations about browser and standards compliance were removed from
the default Wiking footer.  The accessibility statement was moved to the bottom
bar along with the "Powered by Wiking" statement and the footers now contain
only a contact in the form "Contact: <webmaster_address>", where the addres is
replaced by a link to the configured 'webmaster_address' value.  In Wiking CMS,
the footer text is customizable through the "Sytem Texts" page.