Get rid of set_status in the server interface
The request method 'set_status()' was hackish to implement in the wsgi
interface and also generally it was not nice.  A new method 'start_response()'
and a server interface API method 'start_http_response()' replace the original
method 'send_http_header()' (what a misleading name) and the response status
code is passed directly to this new method.  The old method
'send_http_header()' is kept for backwards compatibility (it always sends the
200 cod, which should be ok for its uses outsize wiking).