Fix translation generation
The translations Makefile didn't work since WSGI branch merge as it is not
possible to import the server interface modules (the file
mod_python_interface.py can only be imported in the Apache server
environment).  As the approach with importing all source files to be able to
determine their gettext domain proved to cause lots of troubles (invalid python
path, etc), I decided to abandon this aproach completely.  It should be enough
to distinguish files by their path (such as wiking-cms domain files are located
in lib/wiking/cms/) and the exceptions can be managed manually in the Makefile
(I believe there are none anyway).