Add a codebook of countries
The codebook of countries is currently not used by Wiking CMS itself, but may
be practical for extension applications (currently E-shop).  It is also planned
to make use of it in Wiking CMS for definition of applicable locales.  Locales
are combinations of language/country, such as en-US, en-GB, de-DE, de-AT.
Wiking and Wiking CMS currently ignore countries and only care about languages.
This is quite restricting when it is necessary to care about country specifics.