Add DB support for multisite CMS content
This commit modifies the database to add multisite support, however the
multisite support is not yet implemented in the application code.  One nice
side affect of this change is that all affected DB objects were renamed to a
common prefix cms_ and to make more sense (it was necessary to recreate them
anyway).