Allow chunked response data in WSGI interface
This change involved quite an important API change.  The response data are not
written back to the client through req.write() (the imperative approach), but
they are now handled automatically by the server interface.  The wiking
application simply returns them as an iterable object (which may be a simple
list in simple cases) and a generator in cases where the data need to be
written gradually.  Most applications should not need to be changed, as writing
request data is usually not performed directly by applications.  They just need
to make sure they pass back the return values of methods Request.send_file() or
Request.send_response() and they don't use Request.write().