Completely avoid subselects on referer link generation
Codebook fields which don't define inline_referer nad don't refer by key will
now not have links.  This is the problem of the specification and should be
fixed there.  This commit also adds logging of such problematic codebook
definitions, so it should be now simple to reveal and fix them.