Allow publication publishing/unpublishing and apply edits immediately
The concept of delayed publication of page changes doesn't fit much to
publications.  It doesn't work for the whole publication, but for siggle pages,
which is very confusing.  We need another mechaniosm for whole publication
versioning so we give up the existing support for uncommited page changes and
autocommit all changes to publications and their chapters.