Work around iOS 10 media issue
Without this work around, media playback would not work with iOS 10 for
media files which require authentication.

AppleCoreMedia starting from iOS 10 downloads files referenced in
<audio> and <video> tags without sending all cookies.  Only cookies with
expiration set are sent with the request and others are ignored.

Forcing cookie expiration on iOS 10 devices unfortunately causes the
user sessions to be always persistent in this case, but this actually
doesn't seem to be a huge problem because the browser is typically never
closed on iOS so the users won't expect the session to end unless they
explicitly log out.  Also the iOS devices are often personal so the risk
of stealing user's session is minimized by their typical usage scenario.

At this time it is not known whether Apple is going to fix this issue
and in which iOS version, so the work around is applied for all 10.x
iOS versions.