Drop 'enabled' attribute from 'wiking.Binding'
This attribute is now supported in the parent class in Pytis.  Usage of
wiking.Binding should be now limited only to cases, where the 'form'
attribute must be used and it whould be explicitly referenced as
'wiking.Binding'.  'pytis.presentation.Binding' should be used
everywhere else (and may be imported into the namespeace and used just
as 'Binding').