Improve VO menu accessibility
This makes the submenu items accessible through regular VO navigation.  The role
"tree" was presented as "table, nothing selected" which was impossible to dive
into.  Now the submenu items are navigable even without the need to dive into
the table (which used to be the case on the older menu implementation (prior to
making the menu responsive).