Refactor error handling and reporting
The initial intention was to omit personal information from bug report
emails.  But as the error reporting code was scattered at too many
places, the whole problem required a bit of refactoring.  On the way
down I also noticed that some code keeps references to 'sys.exc_info()'
tracebacks which is discouraged in its documentation, so this also
needed to be reworked.

Finally the Internal Server Error response page content was improved to
give more relevant information and to avoid leaking error messages which
are useless for ordinary users but may be very useful for hackers.