Avoid storing personal information in session history
IP addres and User Agent may be used to uniquely identify someone on the
internet and we need to avoid storing it for identified persons for GDPR
compatibility.  This information, however, may be important in the
history of failed login attempts.  Thus we split the session history and
the failed logins history into two tables, where the later keeps this
information, because it may be considered as anonymous.

Also session cookies were reduced.  Instead of two cookies, one of which
contains the users login (email address), we now have just one cookie
containing completely anonymous information.

The changes also required some refactoring which further allowed some
API cleanup.