Allow current top menu item expansible
The current top menu item was intentionally not expansible on wide screens
because its submenu is already present in the page inline as the left side menu
panel.  This however seems to be confusing when the user doesn't understand
that the left panel replaces the submenu.  The navigatin may seem confusing in
this case, because the menu is sometimes expansible (when not current and also
on narrow screens) and sometimes not (when it becomes current on a wide
screen).

This change leads to more consistency.  The left panel may be understood as
optional navigation element providing more context and simplifying navigation
on wide screens.  It is redundant by nature.  It doesn't influence the main
menu in any way now.