Store original and resized image sizes for CMS attachments
The existing fields width and height were unused and we didn't have fields for
storing the resized size, so add new fields and compute their values as needed.
Also modify the update-thumbnails.py script to update the values correctly.
Existing sites should run the script after updating to this version to update
the database appropriately.