Set aria-hidden on submenu element
The submenu is redundant.  It can be alternatively accessed from the main menu
drop-down.  So it is only confusing for AT users and adds unnecessary cluter to
the user interface.

Setting aria-hidden doesn't exclude the submenu from keyboard nagigation.  This
is good for keyboard users but may be still confusing for screen reader users
if they use standard keyboard navigation.  If they use screen reader
navigation, the submenu is ignored.  It would seem nice to be able to exclude
the submenu also from standard keyboard navigation if we detect screen reader
usage, but this is probably not possible.