Define pyproject.toml for Flit packaging
This step requires moving lib/wiking to project root as Flit doesn't support
package files located in a subdirectory.  The original motivation for the
existence of the lib subdirectory was to be able to include it in PYTHONPATH.
This is actually not necessary anymore when deploying as a package, so moving
wiking to project root seems more than logical.